Refactor JS lib into own files
